py/requirements.txt
neo4j==5.23.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
pyyaml==6.0.2
Faker>=25.0